% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Calculate_TempStressIndex.R
\name{Calculate_TempStressIndex}
\alias{Calculate_TempStressIndex}
\title{Function to calculate temperature stress index for different species on the basis of air temperature}
\usage{
Calculate_TempStressIndex(climdata, speciesdf)
}
\arguments{
\item{climdata}{data frame of climate data}

\item{speciesdf}{data frame of species temperature and precipitation characteristic data}
}
\value{
dataframe of species temperature stress status by year and the number of high temperature stress days that year
}
\description{
Calculate Temperature Stress Index per Species.
Compute climate stress days, or the number of days per year where tmax >86 deg F (>30 deg C).
If species stress days <= climate stress days, than the species stressed column equals yes, otherwise it equals no.
}
\references{
http://agron-www.agron.iastate.edu/courses/Agron541/classes/541/lesson04a/4a.2.html
Air temperatures of >86 deg F (>30 deg C) are considered high stress temperatures for most plant species
https://plants.usda.gov/
Plant characteristic information from the USDA
}
