% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalculateAC_HeatingCost.R
\name{CalculateAC_heatingCost}
\alias{CalculateAC_heatingCost}
\title{Heating and Cooling Cost Function}
\usage{
CalculateAC_heatingCost(climatedata, price_heating_per_degree = 0.8,
  price_cooling_per_degree = 0.8, interactive_plot_selection = T)
}
\arguments{
\item{climatedata}{This is where you put in a climate dataset.}

\item{price_heating_per_degree}{this is the cost of every incremental increasing in heating due to the difference in actual heating with the baseline.}

\item{price_cooling_per_degree}{same as above, but for cooling.}

\item{interactive_plot_selection}{If T, the function returns a plot of the cost datapoints. The AC costs are in blue, the heating costs are in red.}
}
\value{
a list of the dataframe with date, month, year, and the cost of heating and cooling for that day.
}
\description{
This function takes climate data with daily minimum and maximum temperatures and returns the daily cost of air conditions and heating.
}
\details{
Function 5: Heating and Cooling Cost Function

The equation used to calculate the heating and cooling cost goes as the following: We assume the user will turn on heating if the temperature maximum is 8 degrees C. We assume the user will turn on AC if the temperature minimum is 14 degrees. Simply turning on the AC or the heater costs $10, and every additional degree above the baseline (8 C or 14 C) costs a specified price_per_degree. The default is 0.8, but this can be change by manipulating the appropriate param. This means that with every degree increase above or below the baseline temp, the cost goes up 0.8 cents. So, you will spend $10.8 if the average outdoor temp is 15 degrees.
}
\examples{
 CalculateAC_heatingCost(climatedata = clim, price_heating_per_degree = 0.8,
price_cooling_per_degree = 0.8, interactive_plot_selection = T)
}
\author{
Sofie McComb & Margaux Sleckman
}
