% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalculateMonthlyPrecip.R
\name{CalculateMonthlyPrecip}
\alias{CalculateMonthlyPrecip}
\title{Function 2: Calculate Monthly Precipitation}
\usage{
CalculateMonthlyPrecip(clim_data)
}
\arguments{
\item{clim_data}{the climate dataset. Input the clim dataframe from this package}
}
\value{
A list that includes a dataframe of average weekly precipitation per month across all years and a dataframe of average weekly precipitaiton per year
}
\description{
This function takes the climate dataset clim and turns it into an multidimentional array to then output the average weekly precip per month across the years and the average weekly precipitation across the years.
}
\examples{
CalculateMonthlyPrecip(clim_data = clim)
}
\author{
Sofie McComb & Margaux Sleckman
}
