% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clim.R
\docType{data}
\name{clim}
\alias{clim}
\title{Climate Data from Santa Barbara County}
\format{A data frame with 27274 rows and  variables:
\itemize{
\item date Month/Day/Year
\item tmin minimum temperature (C)
\item tmax maximum temperature (C)
\item rain precipitation (mm/day)
\item year
\item month
\item day
\item wy wateryear
}}
\source{
\url{http://sbc.lternet.edu/data/}
}
\usage{
clim
}
\description{
Data from SB-LTER metereology station, daily
meterologic data including minimum temperature,
maximum temperature, precipitation, water year,
and the date
}
\keyword{datasets}
