% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotTempBySeason.R
\name{PlotTempBySeason}
\alias{PlotTempBySeason}
\title{Calculate and Plot Temperature by Season}
\usage{
PlotTempBySeason(climdata)
}
\arguments{
\item{climdata}{Data frame of climate data, including temperature, precipitation, water year, and dates}
}
\value{
List with the following items
\describe{
\item{ClimateDF}{Data frame of maximum and minimum temperature by season over the years}
\item{ClimatePlot}{Plot of maximum and minimum temperature over the years facet-wrapped by season}
}
}
\description{
Function to calculate and plot average maximum and minimum temperature by season across the years.
}
\details{
Seasons were defined as: Spring (March, April, May), Summer (June, July, August),
Fall (September, October, November), and Winter (December, January, February).
Temperature in is degrees Celsius.

Function can be used to assess how average seasonal temperature extremes have changed annually,
which can provide important information for assessing how climate has potentially shifted in the region over time.
}
\author{
Sofie McComb & Margaux Sleckman
}
