% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/species.R
\docType{data}
\name{species}
\alias{species}
\title{Sample Plant Characteristic Data}
\format{A data frame with 3 rows and 7 variables:
\itemize{
\item Species species identifier
\item Scientific species scientific name
\item Common species common name
\item DroughtTolerance drought tolerance capability (low, medium, high)
\item StressDays # of days species can tolerate high temperatures per year before becoming stressed
\item PrecipMin minimum precipitation tolerated by the species
\item PrecipMax maximum precipitation tolerated by the species
}}
\source{
\url{#http://agron-www.agron.iastate.edu/courses/Agron541/classes/541/lesson04a/4a.2.html}

\url{#https://plants.usda.gov/}
}
\usage{
species
}
\description{
Sample dataset for various Santa Barbara plant species providing climate and growth-based plant characteristics.
Species information given by species identifier, scientific name, and common name.
Dataset provides drought tolerance capability and the number of days species can tolerate high heat stress days per year.
Data also reports minimum and maximum amount of precipitation the species can tolerate.
URLs provide source for using >86 degF (>30 degC) as the baseline temperature for stressful temperature levels
and for USDA plant characteristic information.
Species heat stress days derived from drought tolerance level.
}
\keyword{datasets}
