% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Impact_function.R
\name{calculate_heating_AC_cost}
\alias{calculate_heating_AC_cost}
\title{Heating and Cooling Cost Function}
\usage{
calculate_heating_AC_cost(climatedata, price_heating_per_degree = 0.8,
  price_cooling_per_degree = 0.8, interactive_plot_selection = T)
}
\arguments{
\item{climatedata}{This is where you put in a climate dataset.}

\item{price_heating_per_degree}{this is the cost of every incremental increasing in heating due to the difference in actual heating with the baseline.}

\item{price_cooling_per_degree}{same as above, but for cooling.}
}
\value{
a list of the dataframe with date, month, year, and the cost of heating and cooling for that day.
}
\description{
This function takes climate data with daily minimum and maximum temperatures and returns the daily cost of air conditions and heating.
}
\details{
Function 5: Heating and Cooling Cost Function
}
\author{
Sofie McComb & Margaux Sleckman
}
