% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotTempBySeason.R
\name{PlotTempBySeason}
\alias{PlotTempBySeason}
\title{Function to plot maximum and minimum temperature by season across the years}
\usage{
PlotTempBySeason(climdata)
}
\arguments{
\item{climdata}{data frame of climate data}
}
\value{
dataframe and ggplot of maximum and minimum temperature over the years facet wrapped by season
Add more of the description from calcfishrevenue as basis
}
\description{
Plot Temperature by Season
}
