% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotTempBySeason.R
\name{PlotTempBySeason}
\alias{PlotTempBySeason}
\title{Plot Temperature by Season}
\usage{
PlotTempBySeason(climdata)
}
\arguments{
\item{climdata}{Data frame of climate data}
}
\value{
List with the following items
\describe{
\item{ClimateDF}{Data frame of maximum and minimum temperature by season over the years}
\item{ClimatePlot}{Plot of maximum and minimum temperature over the years facet-wrapped by season}
}
}
\description{
Function to calculate and plot maximum and minimum temperature by season across the years.
}
\details{
Seasons were defined as: Spring (March, April, May), Summer (June, July, August),
Fall (September, October, November), and Winter (December, January, February).
}
\examples{
PlotTempBySeason(climdata=clim)

}
\author{
Sofie McComb & Margaux Sleckman
}
